+++ packages/sdk-node/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 4.10.15
+
+### Patch Changes
+
+- Load `jwtKey` from `CLERK_JWT_KEY` env variable ([#1443](https://github.com/clerkinc/javascript/pull/1443)) by [@dimkl](https://github.com/dimkl)
+
+- Updated dependencies [[`2ad7cf390`](https://github.com/clerkinc/javascript/commit/2ad7cf390ba84b8e767ed6fe136800e38356d79c), [`f0b044c47`](https://github.com/clerkinc/javascript/commit/f0b044c475546e96a5995ef16198e60e35e8098f)]:
+  - @clerk/backend@0.24.0
+
+## 4.10.14
+
+### Patch Changes
+
+- Make all 4 keys (legacy and new) optional in authenticateRequest params ([#1437](https://github.com/clerkinc/javascript/pull/1437)) by [@anagstef](https://github.com/anagstef)
+
+- Updated dependencies [[`3fee736c9`](https://github.com/clerkinc/javascript/commit/3fee736c993b0a8fd157d716890810d04e632962), [`ac4e47274`](https://github.com/clerkinc/javascript/commit/ac4e47274afc2ab3a55a78b388a14bed76600402), [`5957a3da6`](https://github.com/clerkinc/javascript/commit/5957a3da68cde3386c741812e2bc03b5519d00e0)]:
+  - @clerk/backend@0.23.7
+
 ## 4.10.13
 
 ### Patch Changes